Refactor tests.
