+++ tests/test_people_api_successful_operations.py
@@ -1,48 +1,6 @@
-import pytest
-
 from app import models
 
 from tests.factories import BookFactory
-
-
-def test_thing_get_fields(client, mock_thing):
-    """
-        Test that only the specified fields are returned
-    """
-    q_params = {"fields[thing]": "name"}
-
-    res = client.get(f"/thing/{mock_thing.id}", query_string=q_params)
-
-    assert res.status_code == 200
-    assert res.get_json()["data"]["id"] == mock_thing.id
-    assert res.get_json()["data"]["attributes"]["name"] == mock_thing.name
-    assert res.get_json()["data"]["attributes"].get("description") is None
-
-
-def test_validate_swagger(client):
-    res = client.get(f"/swagger.json")
-    assert res.status_code == 200
-    swagger_ = res.get_json()
-
-
-@pytest.mark.xfail  # This test might be incorrect!
-def test_delete_no_type_fails(client, db_session, mock_person_with_3_books_read):
-    json = {
-        "data": [
-            {
-                "id": mock_person_with_3_books_read.books_read[0].id,
-            }
-        ],
-    }
-
-    res = client.delete(f"/People/{mock_person_with_3_books_read.id}/books_read/", json=json)
-    assert res.status_code == 403
-
-    person_books_read_list = (
-        db_session.query(models.Book).filter(models.Book.reader_id == mock_person_with_3_books_read.id).all()
-    )
-
-    assert len(person_books_read_list) == 3
 
 
 def test_delete(client, db_session, mock_person_with_3_books_read):
@@ -108,38 +66,13 @@ def test_patch_reader_person(client, db_session, mock_person_with_3_books_read):
     assert person.comment == ""
 
 
-def test_patch_reader_person_fails_for_invalid_id(client, db_session, mock_person_with_3_books_read):
-    data = {
-        "attributes": {"name": "Reader 0 Changed Name", "email": "reader_email0", "dob": "1988-08-09", "comment": ""},
-        "id": "invalid_id",
-        "type": "People",
-    }
-
-    res = client.patch(f"/People/{mock_person_with_3_books_read.id}", json={"data": data})
-    assert res.status_code == 400
-
-
-def test_add_invalid_book_to_reader_person_books_read_list(client, db_session, mock_person_with_3_books_read):
-    newly_read_book = BookFactory.create(name="mock_Read_book")
-
-    data = [{"id": newly_read_book.id}] # no type
-
-    res = client.post(f"/People/{mock_person_with_3_books_read.id}/books_read", json={"data": data})
-    assert res.status_code == 403
-
-    data = [{"id": "invalid id"}] # invalid id
-
-    res = client.post(f"/People/{mock_person_with_3_books_read.id}/books_read", json={"data": data})
-    assert res.status_code == 404
-
-
 def test_add_book_to_reader_person_books_read_list(client, db_session, mock_person_with_3_books_read):
     newly_read_book = BookFactory.create(name="mock_Read_book")
 
     data = [{"id": newly_read_book.id, "type": newly_read_book._s_type}]
 
     res = client.post(f"/People/{mock_person_with_3_books_read.id}/books_read", json={"data": data})
-    assert res.status_code == 204
+    assert res.status_code == 200
 
     person_books_read_list = (
         db_session.query(models.Book).filter(models.Book.reader_id == mock_person_with_3_books_read.id).all()
@@ -159,7 +92,7 @@ def test_get_read_book_from_reader_person_by_id(client, mock_person_with_3_books
     assert response_data["data"]["id"] == newly_read_book.id
 
 
-def test_filter(client):
+def test_people_api_filter(client):
     res = client.get(f"/People/?filter=xx")
     assert res.status_code == 200
 
@@ -169,12 +102,8 @@ def test_filter(client):
     response_data = res.get_json()
     assert response_data["data"][0] is not None
 
-    res = client.get(f"/thing/?filter=xx")
-    assert res.status_code == 200
-    response_data = res.get_json()
 
-
-def test_sort(client):
+def test_people_api_sort(client):
     res = client.get(f"/People/?sort=xx")
     assert res.status_code == 200
 
@@ -182,17 +111,11 @@ def test_sort(client):
     assert res.status_code == 200
     response_data = res.get_json()
     person_test_id = response_data["data"][0]["id"]
-    
+
     res = client.get(f"/People/?sort=-name")
     assert res.status_code == 200
     response_data = res.get_json()
     assert response_data["data"][0]["id"] != person_test_id
-    
-    res = client.get(f"/Publishers/1/books?sort=-title")
-    assert res.status_code == 200
-
-    res = client.get(f"/Publishers/1/books?sort=title")
-    assert res.status_code == 200
 
     res = client.get(f"/People/{person_test_id}/books_read?sort=-title")
     assert res.status_code == 200
@@ -200,11 +123,12 @@ def test_sort(client):
     res = client.get(f"/People/{person_test_id}/books_read?sort=title")
     assert res.status_code == 200
 
-def test_include(client):
+
+def test_people_api_include(client):
     res = client.get(f"/People")
     assert res.status_code == 200
     response_data = res.get_json()
     person_test_id = response_data["data"][0]["id"]
-    
+
     res = client.get(f"/People/{person_test_id}/?include=books_read")
     assert res.status_code == 200
